windows版でウィンドウを閉じたらシステムトレイに行くようにした。アプリ制限もポモドーロも動き続ける。
